requirements.txt
firebase_admin
imdbpy
python-slugify
progress
meilisearch
